requirements.txt
numpy==1.23.5
opencv_python==4.8.0.76
Pillow==9.5.0
scipy==1.10.1
streamlit==1.34.0
tensorflow==2.15
efficientnet==1.1.1
wandb
gdown